refactor: move FeedTopNavFilterState to commons with loadFlowsFor injection
Move FeedTopNavFilterState from amethyst module to commons/commonMain.
The class constructor is simplified to accept a (TopFilter) -> IFeedFlowsType
function instead of all individual dependencies, eliminating the direct
LocalCache.getOrCreateAddressableNote() coupling.

Changes:
- New commons FeedTopNavFilterState with (feedFilterListName, loadFlowsFor, scope)
- New buildLoadFlowsFor() helper in amethyst module encapsulates the
  concrete feed flow construction logic (including LocalCache/ICacheProvider usage)
- Account.kt factory parameter renamed to feedTopNavFilterLoadFlowsFactory,
  now returns (TopFilter) -> IFeedFlowsType; Account constructs FeedTopNavFilterState directly
- Original file replaced with typealias for backward compatibility
- All 3 call sites updated (AccountCacheState, AccountViewModel x2)
