+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/FeedTopNavFilterState.kt
@@ -20,141 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.topNavFeeds
 
-import com.vitorpamplona.amethyst.commons.model.TopFilter
-import com.vitorpamplona.amethyst.commons.model.location.LocationResult
-import com.vitorpamplona.amethyst.commons.model.nip02FollowLists.Kind3FollowListState
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.serverList.MergedFollowListsState
-import com.vitorpamplona.amethyst.model.topNavFeeds.allFollows.AllFollowsFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows.AllUserFollowsFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows.Kind3UserFollowsFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.aroundMe.AroundMeFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.aroundMe.GeohashFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.chess.ChessFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.global.GlobalFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.hashtag.HashtagFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.noteBased.NoteFeedFlow
-import com.vitorpamplona.amethyst.model.topNavFeeds.relay.RelayFeedFlow
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.normalizeRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
-
-class FeedTopNavFilterState(
-    val feedFilterListName: MutableStateFlow<TopFilter>,
-    val kind3Follows: StateFlow<Kind3FollowListState.Kind3Follows>,
-    val allFollows: StateFlow<com.vitorpamplona.amethyst.commons.model.serverList.MergedFollowListsState.AllFollows>,
-    val locationFlow: () -> StateFlow<LocationResult>,
-    val followsRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val blockedRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val proxyRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val relayFeeds: StateFlow<Set<NormalizedRelayUrl>>,
-    val caches: FeedDecryptionCaches,
-    val signer: NostrSigner,
-    val scope: CoroutineScope,
-) {
-    fun loadFlowsFor(listName: TopFilter): IFeedFlowsType =
-        when (listName) {
-            TopFilter.Global -> {
-                GlobalFeedFlow(followsRelays, proxyRelays, relayFeeds)
-            }
-
-            TopFilter.AllFollows -> {
-                AllFollowsFeedFlow(allFollows, followsRelays, blockedRelays, proxyRelays)
-            }
-
-            TopFilter.AllUserFollows -> {
-                AllUserFollowsFeedFlow(allFollows, followsRelays, blockedRelays, proxyRelays)
-            }
-
-            TopFilter.DefaultFollows -> {
-                Kind3UserFollowsFeedFlow(kind3Follows, followsRelays, blockedRelays, proxyRelays)
-            }
-
-            TopFilter.AroundMe -> {
-                AroundMeFeedFlow(locationFlow(), followsRelays, proxyRelays)
-            }
-
-            TopFilter.Chess -> {
-                ChessFeedFlow(followsRelays, proxyRelays)
-            }
-
-            is TopFilter.Community -> {
-                NoteFeedFlow(
-                    LocalCache
-                        .getOrCreateAddressableNote(listName.address)
-                        .flow()
-                        .metadata.stateFlow,
-                    signer,
-                    followsRelays,
-                    blockedRelays,
-                    proxyRelays,
-                    caches,
-                )
-            }
-
-            is TopFilter.PeopleList -> {
-                NoteFeedFlow(
-                    LocalCache
-                        .getOrCreateAddressableNote(listName.address)
-                        .flow()
-                        .metadata.stateFlow,
-                    signer,
-                    followsRelays,
-                    blockedRelays,
-                    proxyRelays,
-                    caches,
-                )
-            }
-
-            is TopFilter.MuteList -> {
-                NoteFeedFlow(
-                    LocalCache
-                        .getOrCreateAddressableNote(listName.address)
-                        .flow()
-                        .metadata.stateFlow,
-                    signer,
-                    followsRelays,
-                    blockedRelays,
-                    proxyRelays,
-                    caches,
-                )
-            }
-
-            is TopFilter.Geohash -> {
-                GeohashFeedFlow(listName.tag, followsRelays, proxyRelays)
-            }
-
-            is TopFilter.Hashtag -> {
-                HashtagFeedFlow(listName.tag, followsRelays, proxyRelays)
-            }
-
-            is TopFilter.Relay -> {
-                RelayFeedFlow(listName.url.normalizeRelayUrl())
-            }
-        }
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val flow: StateFlow<IFeedTopNavFilter> =
-        feedFilterListName
-            .transformLatest { listName ->
-                emitAll(loadFlowsFor(listName).flow())
-            }.onStart {
-                loadFlowsFor(feedFilterListName.value).startValue(this)
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                loadFlowsFor(feedFilterListName.value).startValue(),
-            )
-}
+typealias FeedTopNavFilterState = com.vitorpamplona.amethyst.commons.model.topNavFeeds.FeedTopNavFilterState
